requirements.txt
requests>=2.31
tenacity>=8.2
pyyaml>=6.0
pandas>=2.1
pyarrow>=14.0
pytest>=7.4
